requirements.txt
asgiref==3.2.3
cffi==1.14.0
cryptography==2.8
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
pkg-resources==0.0.0
pycparser==2.19
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
